+++ src/com/intellij/idea/plugin/hybris/impex/codeInsight/hints/ImpExDefaultValueDeclarativeInlayHintsCollector.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -32,7 +32,9 @@ class ImpExDefaultValueDeclarativeInlayHintsCollector : SharedBypassCollector {
         if (element !is ImpexValueGroup) return
         if (element.value != null) return
 
-        val defaultValue = element.computeValue() ?: return
+        val defaultValue = element.computeValue()
+            ?.takeIf { it.isNotBlank() }
+            ?: return
 
         sink.addPresentation(
             position = InlineInlayPosition(element.textRange.startOffset + 1, true),